refactor(importer): Moved persistense directory into infrastructure
