+++ module/importer/src/Infrastructure/Persistence/Query/DbalImportQuery.php
@@ -7,16 +7,16 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Importer\Persistence\Dbal\Query;
+namespace Ergonode\Importer\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Query\QueryBuilder;
 use Ergonode\Core\Domain\ValueObject\Language;
 use Ergonode\Grid\DataSetInterface;
 use Ergonode\Grid\DbalDataSet;
-use Ergonode\SharedKernel\Domain\Aggregate\ImportId;
-use Ergonode\SharedKernel\Domain\Aggregate\ImportErrorId;
 use Ergonode\Importer\Domain\Query\ImportQueryInterface;
+use Ergonode\SharedKernel\Domain\Aggregate\ImportErrorId;
+use Ergonode\SharedKernel\Domain\Aggregate\ImportId;
 use Ergonode\SharedKernel\Domain\Aggregate\SourceId;
 use Symfony\Contracts\Translation\TranslatorInterface;
