+++ module/importer/src/Infrastructure/Persistence/Repository/Mapper/DbalImportMapper.php
@@ -7,13 +7,13 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Importer\Persistence\Dbal\Repository\Mapper;
+namespace Ergonode\Importer\Infrastructure\Persistence\Repository\Mapper;
 
 use Ergonode\Importer\Domain\Entity\Import;
 
 /**
  */
-class ImportMapper
+class DbalImportMapper
 {
     /**
      * @param Import $import
