+++ module/importer/src/Infrastructure/Persistence/Repository/Mapper/DbalSourceMapper.php
@@ -7,14 +7,14 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Importer\Persistence\Dbal\Repository\Mapper;
+namespace Ergonode\Importer\Infrastructure\Persistence\Repository\Mapper;
 
 use Ergonode\Importer\Domain\Entity\Source\AbstractSource;
 use JMS\Serializer\SerializerInterface;
 
 /**
  */
-class SourceMapper
+class DbalSourceMapper
 {
     /**
      * @var SerializerInterface
