+++ module/importer/src/Resources/config/services.yml
@@ -3,12 +3,9 @@ parameters:
 
 services:
     _defaults:
-            autowire: true
-            autoconfigure: true
-            public: false
-
-    Ergonode\Importer\Persistence\:
-        resource: '../../Persistence/*'
+        autowire: true
+        autoconfigure: true
+        public: false
 
     Ergonode\Importer\Application\:
         resource: '../../Application/*'
@@ -36,9 +33,9 @@ services:
         resource: '../../Infrastructure/JMS/Serializer/Handler/*'
         tags: ['jms_serializer.subscribing_handler']
 
-    Ergonode\Importer\Domain\Repository\ImportRepositoryInterface: '@Ergonode\Importer\Persistence\Dbal\Repository\DbalImportRepository'
-    Ergonode\Importer\Domain\Repository\ImportErrorRepositoryInterface: '@Ergonode\Importer\Persistence\Dbal\Repository\DbalImportErrorRepository'
-    Ergonode\Importer\Domain\Query\ImportQueryInterface: '@Ergonode\Importer\Persistence\Dbal\Query\DbalImportQuery'
+    Ergonode\Importer\Domain\Repository\ImportRepositoryInterface: '@Ergonode\Importer\Infrastructure\Persistence\Repository\DbalImportRepository'
+    Ergonode\Importer\Domain\Repository\ImportErrorRepositoryInterface: '@Ergonode\Importer\Infrastructure\Persistence\Repository\DbalImportErrorRepository'
+    Ergonode\Importer\Domain\Query\ImportQueryInterface: '@Ergonode\Importer\Infrastructure\Persistence\Query\DbalImportQuery'
 
     Ergonode\Importer\Application\Service\Upload\UploadService:
         arguments:
